Enable citations to be given in subject form.
`\cite` is not intended to be used in this way; it may merely be a short numerical identifyer (possibly in superscript). For refering to a work or its authors in grammatical form, like: "John Doe [Doe87] claimed...", the `\textcite` command is intended. Will can embed a change to this command into the resolver mechanism of the DOIs.