Use Measure for the width of a multirow cell
The width of a multirow cell can be a dimension or '*' (indicating
that the natural width of the text is to be used) or '=' (indicating
that the specified width of the column in which the multirow entry is
set should be used).

Because of those two special cases its type has initially been set as
'LaTeXC l => l'. But it can be 'Measure', which supports the special
cases as 'CustomMeasure' and better fits the purpose of the width
argument.