Add new module for maths commands that don't require AMS.
This forms conceptually part of the `Base.Commands` module,
but is kept separate because
a) that module is very big already
b) these commands should also be exported from the AMSMath module,
if only for the sake of compatibility with HaTeX versions up
until now where they were _defined_ in that module.
It is more elegant to just re-export a complete module, rather
than a bunch of individual identifiers.