Make DOI references usable in a more general way.
Previously, `documentWithDOIReferences` was the only way to use content
containing `citeDOI` etc., but it only works for a very restricted use
case (when the entire document is assembled in the same monadic context,
and not e.g. used within an existing LaTeX template).

The new `applyDOIReferenceResolves` is more difficult to use, but also
much more general.